requirements.txt
pygrib
cdo==1.7.0
# Compile cmor source and run python setup.py install
# CMOR==3.3.0
f90nml==0.20
netCDF4==1.2.7
nose==1.3.7
numpy==1.11.3
python-dateutil==2.6.0
six==1.10.0
testfixtures==4.13.3
xlrd==1.0.0
XlsxWriter==1.0.2
